Conflicts:
	.gitignore

+++ couchpotato/core/_base/updater/main.py
@@ -20,6 +20,8 @@ log = CPLog(__name__)
 
 class Updater(Plugin):
 
+    available_notified = False
+
     def __init__(self):
 
         if Env.get('desktop'):
@@ -30,7 +32,7 @@ class Updater(Plugin):
             self.updater = SourceUpdater()
 
         fireEvent('schedule.interval', 'updater.check', self.autoUpdate, hours = 6)
-        addEvent('app.load', self.check)
+        addEvent('app.load', self.autoUpdate)
         addEvent('updater.info', self.info)
 
         addApiView('updater.info', self.getInfo, docs = {
@@ -51,15 +53,31 @@ class Updater(Plugin):
 
     def autoUpdate(self):
         if self.check() and self.conf('automatic') and not self.updater.update_failed:
-            self.updater.doUpdate()
+            if self.updater.doUpdate():
+
+                # Notify before restarting
+                try:
+                    if self.conf('notification'):
+                        info = self.updater.info()
+                        version_date = datetime.fromtimestamp(info['update_version']['date'])
+                        fireEvent('updater.updated', 'Updated to a new version with hash "%s", this version is from %s' % (info['update_version']['hash'], version_date), data = info)
+                except:
+                    log.error('Failed notifying for update: %s', traceback.format_exc())
+
+                fireEventAsync('app.restart')
+
+                return True
+
+        return False
 
     def check(self):
         if self.isDisabled():
             return
 
         if self.updater.check():
-            if self.conf('notification') and not self.conf('automatic'):
+            if not self.available_notified and self.conf('notification') and not self.conf('automatic'):
                 fireEvent('updater.available', message = 'A new update is available', data = self.updater.info())
+                self.available_notified = True
             return True
 
         return False
@@ -157,17 +175,10 @@ class GitUpdater(BaseUpdater):
             self.repo.saveStash()
 
             log.info('Updating to latest version')
-            info = self.info()
             self.repo.pull()
 
             # Delete leftover .pyc files
             self.deletePyc()
-
-            # Notify before returning and restarting
-            version_date = datetime.fromtimestamp(info['update_version']['date'])
-            fireEvent('updater.updated', 'Updated to a new version with hash "%s", this version is from %s' % (info['update_version']['hash'], version_date), data = info)
-
-            fireEventAsync('app.restart')
 
             return True
         except:
@@ -197,7 +208,7 @@ class GitUpdater(BaseUpdater):
     def check(self):
 
         if self.update_version:
-            return
+            return True
 
         log.info('Checking for new version on github for %s', self.repo_name)
         if not Env.get('dev'):
@@ -254,15 +265,13 @@ class SourceUpdater(BaseUpdater):
             tar.close()
             os.remove(destination)
 
-            self.replaceWith(os.path.join(extracted_path, os.listdir(extracted_path)[0]))
-            self.removeDir(extracted_path)
+            if self.replaceWith(os.path.join(extracted_path, os.listdir(extracted_path)[0])):
+                self.removeDir(extracted_path)
 
-            # Write update version to file
-            self.createFile(self.version_file, json.dumps(self.update_version))
+                # Write update version to file
+                self.createFile(self.version_file, json.dumps(self.update_version))
 
-            fireEventAsync('app.restart')
-
-            return True
+                return True
         except:
             log.error('Failed updating: %s', traceback.format_exc())
 
@@ -273,7 +282,7 @@ class SourceUpdater(BaseUpdater):
         app_dir = ss(Env.get('app_dir'))
 
         # Get list of files we want to overwrite
-        self.deletePyc(only_excess = False)
+        self.deletePyc()
         existing_files = []
         for root, subfiles, filenames in os.walk(app_dir):
             for filename in filenames:
@@ -286,18 +295,30 @@ class SourceUpdater(BaseUpdater):
 
                 if not Env.get('dev'):
                     try:
-                        os.remove(tofile)
-                    except:
-                        pass
+                        if os.path.isfile(tofile):
+                            os.remove(tofile)
 
-                    try:
-                        os.renames(fromfile, tofile)
+                        dirname = os.path.dirname(tofile)
+                        if not os.path.isdir(dirname):
+                            self.makeDir(dirname)
+
+                        os.rename(fromfile, tofile)
                         try:
                             existing_files.remove(tofile)
                         except ValueError:
                             pass
-                    except Exception, e:
-                        log.error('Failed overwriting file: %s', e)
+                    except:
+                        log.error('Failed overwriting file "%s": %s', (tofile, traceback.format_exc()))
+                        return False
+
+        if Env.get('app_dir') not in Env.get('data_dir'):
+            for still_exists in existing_files:
+                try:
+                    os.remove(still_exists)
+                except:
+                    log.error('Failed removing non-used file: %s', traceback.format_exc())
+
+        return True
 
 
     def removeDir(self, path):
